+++ CHANGES.md
@@ -1,3 +1,6 @@
+* 0.1.4
+    * New `write` command that writes the buffer to disk
+    * BUGFIX: newline chars didn't render if they were the last instructions
 * 0.1.3
     * BUGFIX: instructions per second did not respect `sleep` instructions
 * 0.1.2